pyproject.toml
[project]
name = "flypush"
version = "0.1.0"
description = "A SaaS web application for managing Drosophila stocks in research laboratories"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Giorgio Gilestro", email = "giorgio@gilest.ro"}
]
keywords = ["drosophila", "fly stocks", "laboratory", "research", "management"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pymysql>=1.1.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "qrcode[pil]>=7.4.0",
    "python-barcode>=0.15.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | alembic/versions
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".git", ".venv", "__pycache__", "alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",  # Line too long (handled by black)
    "B008",  # Do not perform function calls in argument defaults (FastAPI Depends() pattern)
    "B904",  # Exception chaining - can be added later for better error context
]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]  # Module level import not at top of file (needed for logging setup)
"tests/conftest.py" = ["E402"]  # Module level import not at top of file (needed for env setup)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [".venv", "tests", "alembic/versions"]
skips = ["B101"]  # Skip assert_used check as it's used in tests
